split errors package
