+++ errors/error.go
@@ -1,4 +1,4 @@
-package main
+package errors
 
 import (
 	"fmt"
